Update README.md
The brain is the foundation of the human body and is responsible for all functions. Neurons are in charge of our bodies' responses. In the same way, artificial neurons that function similarly to biological brains are made. The response of the artificial neurons is obtained in this research by varying the threshold values and activation functions of logic gates. Every neuron model is made up of a processing element with a single input and a synaptic input connection.
FOR OR:
The OR gate is a type of digital logic gate that implements logical disjunction—it operates in accordance with the truth table. If one or both of the gate's inputs are high, the output is 1. If both inputs are low, the outcome is low (0). 
A plus (+) sign is used to indicate the or operation.
FOR AND:
It is a logic gate that implements conjunction.
When both inputs are high, the output will be high (1); otherwise, it will be low (0).
FOR XOR:
It is also known as an XOR gate, an exclusive gate, or a gate. When the number of true inputs is odd, it produces a true output. If both inputs are true and both outputs are false, the output is false. 
These are used in computers to implement binary addition.